+++ tools/hcollector.cxx
@@ -95,7 +95,7 @@ int HCollector::send_line( char const* line_ )
 	int long ctr = 0;
 	int cRC = 0;
 	int error = -1;
-	int long length = ::strlen( line_ );
+	int length = static_cast<int>( ::strlen( line_ ) );
 	HString line, localCopy;
 	if ( length < 1 )
 		return ( 0 );
@@ -110,7 +110,7 @@ int HCollector::send_line( char const* line_ )
 	line.format ( "%s%02x%02x%s\n", PROTOCOL::DTA,
 			length & 0x0ff, cRC & 0x0ff, localCopy.raw() );
 	::memset( _readBuf, 0, PROTOCOL::RECV_BUF_SIZE );
-	length += ::strlen( PROTOCOL::DTA );
+	length += static_cast<int>( ::strlen( PROTOCOL::DTA ) );
 	length += ( 2 /* for lenght */ + 2 /* for crc */ + 1 /* for newline */ );
 	while ( ::strncmp( _readBuf, PROTOCOL::ACK, ::strlen( PROTOCOL::ACK ) ) )
 		{
@@ -133,8 +133,8 @@ int HCollector::receive_line( HString& line_ )
 	int ctr = 0;
 	int cRC = 0, pCRC = -1;
 	int long length = 0, pLength = -1;
-	int ackLenght = ::strlen( PROTOCOL::ACK );
-	int errLenght = ::strlen( PROTOCOL::ERR );
+	int ackLenght = static_cast<int>( ::strlen( PROTOCOL::ACK ) );
+	int errLenght = static_cast<int>( ::strlen( PROTOCOL::ERR ) );
 	/* P prefix means sender transmission side data */
 	while ( ( pCRC != cRC ) || ( pLength != length ) )
 		{
@@ -148,7 +148,7 @@ int HCollector::receive_line( HString& line_ )
 			}
 		flush( TCIFLUSH );
 		line_ = _line;
-		line_.shift_left(	::strlen( PROTOCOL::DTA ) + 2 /* for lenght */ + 2 /* for crc */ );
+		line_.shift_left(	static_cast<int long>( ::strlen( PROTOCOL::DTA ) ) + 2 /* for lenght */ + 2 /* for crc */ );
 		length = line_.get_length() - 1;
 		for ( ctr = 0; ctr < length; ctr ++ )
 			cRC += line_[ ctr ];
@@ -182,7 +182,7 @@ int HCollector::establish_connection ( int timeOut_ )
 	 Either waiting part runs before initializing part (the easy case),
 	 or initializing part runs before waiting part (here comes the trouble).
 */
-	int lenght = ::strlen( PROTOCOL::SYN ), error = -1;
+	int lenght = static_cast<int>( ::strlen( PROTOCOL::SYN ) ), error = -1;
 	if ( _fileDescriptor < 0 )
 		M_THROW( _error_, _fileDescriptor );
 	::memset( _readBuf, 0, PROTOCOL::RECV_BUF_SIZE );
@@ -210,7 +210,7 @@ int HCollector::wait_for_connection ( int timeOut_ )
 	{
 	M_PROLOG
 	int error = - 1;
-	int lenght = ::strlen( PROTOCOL::ACK );
+	int lenght = static_cast<int>( ::strlen( PROTOCOL::ACK ) );
 	if ( _fileDescriptor < 0 )
 		M_THROW ( _error_, _fileDescriptor );
 	::memset( _readBuf, 0, PROTOCOL::RECV_BUF_SIZE );
